+++ Common/src/main/java/fuzs/arcanelanterns/world/level/block/BrilliantLanternBlock.java
@@ -1,7 +1,7 @@
 package fuzs.arcanelanterns.world.level.block;
 
 import fuzs.arcanelanterns.init.ModRegistry;
-import fuzs.arcanelanterns.world.level.block.entity.BriliantLanternBlockEntity;
+import fuzs.arcanelanterns.world.level.block.entity.BrilliantLanternBlockEntity;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntity;
@@ -10,20 +10,20 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
 
-public class BriliantLanternBlock extends LanternEntityBlock {
+public class BrilliantLanternBlock extends LanternEntityBlock {
 
-    public BriliantLanternBlock(Properties properties) {
+    public BrilliantLanternBlock(Properties properties) {
         super(properties);
     }
 
     @Override
     public @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType) {
-        return !level.isClientSide ? createTickerHelper(blockEntityType, ModRegistry.BRILIANT_LANTERN_BLOCK_ENTITY.get(), BriliantLanternBlockEntity::tick) : null;
+        return !level.isClientSide ? createTickerHelper(blockEntityType, ModRegistry.BRILLIANT_LANTERN_BLOCK_ENTITY.get(), BrilliantLanternBlockEntity::tick) : null;
     }
 
     @Nullable
     @Override
     public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
-        return new BriliantLanternBlockEntity(pos, state);
+        return new BrilliantLanternBlockEntity(pos, state);
     }
 }