+++ Common/src/main/java/fuzs/arcanelanterns/world/level/block/SparkBlock.java
@@ -1,20 +1,25 @@
 package fuzs.arcanelanterns.world.level.block;
 
+import fuzs.arcanelanterns.init.ModRegistry;
+import fuzs.arcanelanterns.world.level.block.entity.SparkBlockEntity;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleTypes;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.RandomSource;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.*;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.shapes.CollisionContext;
-import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
+import org.jetbrains.annotations.Nullable;
 
 @SuppressWarnings("deprecation")
-public class SparkBlock extends Block {
-    private static final VoxelShape SHAPE = Shapes.box(0.4, 0.4, 0.4, 0.6, 0.6, 0.6);
+public class SparkBlock extends BaseEntityBlock {
+    private static final VoxelShape SHAPE = Block.box(6.0, 6.0, 6.0, 10.0, 10.0, 10.0);
 
     public SparkBlock(BlockBehaviour.Properties properties) {
         super(properties);
@@ -26,9 +31,49 @@ public class SparkBlock extends Block {
     }
 
     @Override
-    public void animateTick(BlockState state, Level world, BlockPos pos, RandomSource random) {
-        if (world.isClientSide) {
-            world.addParticle(ParticleTypes.FLAME, pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5, 0, 0.03, 0);
+    public RenderShape getRenderShape(BlockState state) {
+        return RenderShape.MODEL;
+    }
+
+    @Override
+    public boolean isRandomlyTicking(BlockState state) {
+        return true;
+    }
+
+    @Override
+    public void randomTick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random) {
+        if (level.getBlockEntity(pos) instanceof SparkBlockEntity blockEntity) {
+            if (level.getBlockState(blockEntity.pos).is(ModRegistry.FERAL_LANTERN_BLOCK.get())) {
+                return;
+            }
         }
+        level.setBlock(pos, Blocks.AIR.defaultBlockState(), 2);
+    }
+
+    @Override
+    protected void spawnDestroyParticles(Level level, Player player, BlockPos pos, BlockState state) {
+
+    }
+
+    @Override
+    public void playerWillDestroy(Level level, BlockPos pos, BlockState state, Player player) {
+        if (!level.isClientSide()) {
+            level.levelEvent(null, LevelEvent.SOUND_EXTINGUISH_FIRE, pos, 0);
+        }
+
+        super.playerWillDestroy(level, pos, state, player);
+    }
+
+    @Override
+    public void animateTick(BlockState state, Level level, BlockPos pos, RandomSource random) {
+        if (level.isClientSide()) {
+            level.addParticle(ParticleTypes.FLAME, pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5, 0, 0.03, 0);
+        }
+    }
+
+    @Nullable
+    @Override
+    public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
+        return new SparkBlockEntity(pos, state);
     }
 }