+++ Common/src/main/java/fuzs/arcanelanterns/world/level/block/entity/FeralLanternBlockEntity.java
@@ -4,7 +4,7 @@ import fuzs.arcanelanterns.ArcaneLanterns;
 import fuzs.arcanelanterns.config.ServerConfig;
 import fuzs.arcanelanterns.init.ModRegistry;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Vec3i;
+import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
@@ -13,7 +13,6 @@ public class FeralLanternBlockEntity extends LanternBlockEntity {
     private static final String TAG_FLARES = "PlacedFlares";
 
     private int placedFlares;
-    private boolean placeAttempt;
 
     public FeralLanternBlockEntity(BlockPos pos, BlockState state) {
         super(ModRegistry.FERAL_LANTERN_BLOCK_ENTITY.get(), pos, state);
@@ -21,29 +20,39 @@ public class FeralLanternBlockEntity extends LanternBlockEntity {
 
     public static void tick(Level level, BlockPos pos, BlockState state, FeralLanternBlockEntity blockEntity) {
         ServerConfig.FeralLanternConfig config = ArcaneLanterns.CONFIG.get(ServerConfig.class).feralLantern;
-        if (++blockEntity.count > config.delay && !blockEntity.placeAttempt) {
-            final int horizontalRange = config.horizontalRange;
-            final int verticalRange = config.verticalRange;
-            if (blockEntity.placedFlares >= config.maxPlacedFlares) {
-                level.destroyBlock(pos, false);
-                return;
+        if (++blockEntity.count > config.delay && !blockEntity.isDonePlacing()) {
+            BlockPos.MutableBlockPos mutable = pos.mutable();
+            mutable.move(-config.horizontalRange, -config.verticalRange, -config.horizontalRange);
+            mutable.move(level.random.nextInt(config.horizontalRange * 2), level.random.nextInt(config.verticalRange * 2), level.random.nextInt(config.horizontalRange * 2));
+            // max manhattan distance approximation
+            int maxDistance = 5 * (config.horizontalRange + config.verticalRange) / 7;
+            while (mutable.closerThan(pos, maxDistance) && !level.isOutsideBuildHeight(mutable) && level.getBlockState(mutable).getCollisionShape(level, mutable).isEmpty()) {
+                mutable.move(Direction.DOWN);
             }
-            BlockPos check = pos.subtract(new Vec3i(horizontalRange, verticalRange, horizontalRange));
-            check = check.offset(level.random.nextInt(horizontalRange * 2), -level.random.nextInt(verticalRange * 2), level.random.nextInt(horizontalRange * 2));
-            while (check.closerThan(pos, 200)) {
-                blockEntity.placeAttempt = true;
-                if (level.getBlockState(check).isAir() && !(level.getBlockState(check.below()).isAir()) && level.getMaxLocalRawBrightness(check) < config.maxLightLevel) {
-                    level.setBlockAndUpdate(check, ModRegistry.SPARK_BLOCK.get().defaultBlockState());
-                    blockEntity.placeAttempt = false;
-                    blockEntity.placedFlares++;
+            while (mutable.closerThan(pos, maxDistance) && !level.isOutsideBuildHeight(mutable) && !level.getBlockState(mutable).getCollisionShape(level, mutable).isEmpty()) {
+                mutable.move(Direction.UP);
+            }
+            if (level.getMaxLocalRawBrightness(mutable) < config.maxLightLevel) {
+                if (!level.getBlockState(mutable.below()).getCollisionShape(level, mutable.below()).isEmpty()) {
+                    mutable.move(Direction.UP, 3);
+                    for (int i = 0; i < 3 && mutable.closerThan(pos, maxDistance) && !level.getBlockState(mutable).isAir(); i++) {
+                        mutable.move(Direction.DOWN);
+                    }
+                    if (level.getBlockState(mutable).isAir()) {
+                        level.setBlockAndUpdate(mutable, ModRegistry.SPARK_BLOCK.get().defaultBlockState());
+                        if (level.getBlockEntity(mutable) instanceof SparkBlockEntity sparkBlockEntity) {
+                            sparkBlockEntity.pos = pos;
+                        }
+                        blockEntity.placedFlares++;
+                    }
                 }
-                check = check.subtract(new Vec3i(0, 1, 0));
-            }
-            if (!check.closerThan(pos, 100)) {
-                blockEntity.placeAttempt = false;
             }
             blockEntity.count = 0;
         }
+    }
+
+    public boolean isDonePlacing() {
+        return this.placedFlares >= ArcaneLanterns.CONFIG.get(ServerConfig.class).feralLantern.maxPlacedFlares;
     }
 
     @Override