+++ Forge/src/main/java/fuzs/arcanelanterns/data/ModBlockLootProvider.java
@@ -3,6 +3,7 @@ package fuzs.arcanelanterns.data;
 import fuzs.arcanelanterns.init.ModRegistry;
 import fuzs.puzzleslib.api.data.v1.AbstractLootProvider;
 import net.minecraft.data.PackOutput;
+import net.minecraft.world.level.block.Blocks;
 
 public class ModBlockLootProvider extends AbstractLootProvider.Blocks {
 
@@ -14,7 +15,7 @@ public class ModBlockLootProvider extends AbstractLootProvider.Blocks {
     public void generate() {
         this.dropSelf(ModRegistry.LANTERN_MAKER_BLOCK.get());
         this.dropSelf(ModRegistry.LIFE_LANTERN_BLOCK.get());
-        this.dropSelf(ModRegistry.FERAL_LANTERN_BLOCK.get());
+        this.dropOther(ModRegistry.FERAL_LANTERN_BLOCK.get(), Blocks.LANTERN);
         this.dropSelf(ModRegistry.LOVE_LANTERN_BLOCK.get());
         this.dropSelf(ModRegistry.WAILING_LANTERN_BLOCK.get());
         this.dropSelf(ModRegistry.BOREAL_LANTERN_BLOCK.get());